Change: Rename test / plugin to match its counterpart / class name.
